Created abstract class Projectile. This will create a projectile that will behave however you want when you implement it. Currently the only projectile is the KuhlBird, which will run in a straight line until it expires.
