+++ src/main/java/com/huskygames/rekhid/slugger/actor/Projectile.java
@@ -1,11 +1,6 @@
-package com.huskygames.rekhid.actor;
+package com.huskygames.rekhid.slugger.actor;
 
-import com.huskygames.rekhid.Definitions;
 import com.huskygames.rekhid.Rekhid;
-import com.huskygames.rekhid.slugger.actor.Actor;
-import com.huskygames.rekhid.slugger.actor.ActorCircle;
-import com.huskygames.rekhid.slugger.actor.Attack;
-import com.huskygames.rekhid.slugger.actor.Fighter;
 import com.huskygames.rekhid.slugger.resource.Resource;
 import com.huskygames.rekhid.slugger.resource.sprite.SpriteSheet;
 import com.huskygames.rekhid.slugger.util.DoublePair;
@@ -21,12 +16,12 @@ import java.util.Set;
  * Projectile demo class, can be extended and overridden to do different things as need be.
  * Created by Kyle on 11/5/2016.
  */
-public class Projectile extends Fighter {
+public abstract class Projectile extends Fighter {
 
     private Projectiles proj;
     private SpriteSheet sprite;
-    private boolean facingleft;
     private Set<Shape> colliders;
+    private boolean facingleft;
     private Fighter parent;
     private int TTL; // ticks to live
 
@@ -44,19 +39,13 @@ public class Projectile extends Fighter {
         this.sprite = (SpriteSheet) Rekhid.getInstance().getResourceManager()
                 .requestResource(Resource.TEST_PROJECTILE);
 
-        facingleft = vel.getX() < 0 ? true : false;
-        int multiplier = facingleft ? -1 : 1;
-        DoublePair [][] offsets = new DoublePair [][]{new DoublePair[]{new DoublePair(0.0, 5.0)}};
-        DoublePair [][] directions = new DoublePair [][]{new DoublePair[]{new DoublePair(multiplier * 20, -5)}};
-        int [] ticks = new int[]{20};
-        int [][] damages = new int[][]{new int[]{30}};
-        int [][] areas = new int[][]{new int[]{10}};
-
         colliders = new HashSet<Shape>();
-        colliders.add(new ActorCircle(new DoublePair(0.0, 5.0), this, 6.0));
-        theAttack = new Attack(offsets, ticks, damages, directions, areas, this);
 
         logger.info("I've been created!!");
+    }
+
+    public void addCollider(ActorCircle circle){
+        colliders.add(circle);
     }
 
     public int getHeight() {
@@ -66,15 +55,6 @@ public class Projectile extends Fighter {
     @Override
     public Set<Shape> getCollisions() {
         return colliders;
-    }
-
-    @Override
-    public void tick() {
-        TTL--;
-        if(TTL == 0 && parent instanceof Fighter){
-            ((Fighter)parent).removeProjectile();
-            return;
-        }
     }
 
     @Override
@@ -89,8 +69,20 @@ public class Projectile extends Fighter {
 
     public Fighter getParent(){return parent;}
 
+    public boolean decrementTime(){
+        TTL--;
+        if(TTL == 0){
+            return true;
+        }
+        return false;
+    }
+
     @Override
     public String getName() {
         return proj.toString();
+    }
+
+    public boolean isLeft(){
+        return facingleft;
     }
 }
