+++ src/main/java/com/huskygames/rekhid/slugger/actor/Projectiles.java
@@ -1,4 +1,4 @@
-package com.huskygames.rekhid.actor;
+package com.huskygames.rekhid.slugger.actor;
 
 /**
  * Created by Kyle on 11/12/2016.
